+++ s/auth/tokens/claim.ts
@@ -24,18 +24,17 @@ export class Claim<C> {
 		return Date.now() > this.expiresAt
 	}
 
-	static decode<C>(proof: Proof, claimToken: string) {
-		return new this(
-			proof,
-			claimToken,
-			Token.decode<ClaimPayload<C>>(claimToken).payload,
-		)
+	static decode<C>(claimToken: string) {
+		return Token.decode<ClaimPayload<C>>(claimToken)
 	}
 
 	static async verify<C>(proof: Proof, claimToken: string, options: VerificationOptions = {}) {
-		const claim = this.decode<C>(proof, claimToken)
+		const {payload} = this.decode<C>(claimToken)
+		const claim = new this(proof, claimToken, payload)
+
 		if (claim.thumbprint !== proof.thumbprint)
 			throw new Error(`thumbprint mismatch between claim and proof`)
+
 		const loginPubkey = await proof.getLoginPubkey()
 		await loginPubkey.verify(claimToken, options)
 		return claim